Move HampelFilter
