+++ aeon/forecasting/compose/tests/test_pipeline.py
@@ -19,10 +19,10 @@ from aeon.testing.data_generation._legacy import get_examples, make_series
 from aeon.testing.mock_estimators import MockForecaster, MockTransformer
 from aeon.testing.utils.estimator_checks import _assert_array_almost_equal
 from aeon.transformations._legacy.adapt import TabularToSeriesAdaptor
+from aeon.transformations._legacy.outlier_detection import HampelFilter
 from aeon.transformations.detrend import Detrender
 from aeon.transformations.hierarchical.aggregate import Aggregator
 from aeon.transformations.impute import Imputer
-from aeon.transformations.outlier_detection import HampelFilter
 from aeon.utils.index_functions import get_window
 from aeon.utils.validation._dependencies import _check_soft_dependencies
